Remove console.logs

+++ server/controllers/account/AccountController.js
@@ -5,7 +5,11 @@ const jwt = require("jsonwebtoken");
 const User = require("../../models/UserModel");
 const Item = require("../../models/ItemModel");
 const Store = require("../../models/StoreModel");
-const { validateRegister, validateLogin } = require("./validation");
+const {
+  validateRegister,
+  validateLogin,
+  validatePassword,
+} = require("./validation");
 const verifyToken = require("../../utils/verifyToken");
 
 //@routes POST api/account/register
@@ -23,7 +27,6 @@ router.post("/register", validateRegister, async (req, res) => {
       res.status(200).json({ token, userId: user._id });
     })
     .catch((err) => {
-      console.log("ERR", err);
       res.status(400).json({ error: err.message });
     });
 });
@@ -37,13 +40,27 @@ router.post("/login", validateLogin, (req, res) => {
 
 //@routes DELETE api/account/
 //@desc Delete a user and all stores/items
-router.delete("/", verifyToken, (req, res) => {
+router.delete("/", verifyToken, async (req, res) => {
+  const user = await User.findById(req.user.id);
+
+  const validPassword = await validatePassword(
+    req.body.password,
+    user.password
+  );
+
+  if (!validPassword) {
+    res.status(400).json({ error: "Invalid email or password" });
+    return;
+  }
+
   User.findByIdAndDelete(req.user.id).catch(() => {
     res.status(404).json({ error: "User could not be found" });
   });
+
   Store.deleteMany({ userId: req.user.id }).catch(() =>
     res.status(404).json({ error: "User stores could not be deleted" })
   );
+
   Item.deleteMany({ userId: req.user.id })
     .then(() => res.status(204).json({}))
     .catch(() =>
